backend/requirements.txt
Django==5.0.7
djangorestframework
psycopg2-binary
django-cors-headers
camelot-py[cv]
PyMuPDF
gunicorn
gradio-client
